v1.3 Enhancements
* Added new custom header that is a city skyline
* Added new custom header that is a desk with a briefcase on it
* Remove content-header.php as the parent theme (Flat Bootstrap v1.4)
now handles the logic needed in this child theme. So no need to
override it here anymore!
* Removed header.php so that the one from the parent theme is used.
Created a function to filter xsbf_navbar to right-justify the menu
items and replace O's in the site name with a red dot.
* Move the Page Top widget area to below the header image as this looks
better.
* Updated the screenshot.
* Bumped version numbers.
* Includes all the latest enhancements from the parent theme, such as
support for portfolios and testimonials, new page templates, full-width
embedded videos, expanded color palette, and much more.